prci: split ClockGroupDriver into its own file
